Split Streaming API tutorials into two
The Streaming API tutorial has gotten long, so this commit split it into two.
